% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer of the state number to map}

\item{year}{An integer of the year to map}
}
\value{
A plot of the given state and the points of accidents for the given 
year
}
\description{
This function maps the state and pinpoints the accident in the map of the
state and year specified. Returns and error if the state number is invalid
and a message if there are no accidents.
}
\examples{
\dontrun{fars_map_state(1, 2013)}

}
