% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Reader Fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path of the file to read}
}
\value{
This function returns the data table for the data read
}
\description{
This is a function to read the data given a filename and return a datatable
containing it. If the filename does not exist gives an error.
}
\examples{
\dontrun{data <- fars_read("accident_2013.csv")}
\dontrun{data <- fars_read(make_filename(2013))}

}
