% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Multiple year reader}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years to read}
}
\value{
List of tables with the data read for each specified year
}
\description{
This function reads the files of the specified years and returns an error if 
the year file does not exist.
}
\examples{
\dontrun{data <- fars_read_years(c(2013, 2014))}

}
