% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years to summarise}
}
\value{
A summary for the different specified years
}
\description{
This function creates an overview of the different years specified.
}
\examples{
\dontrun{data <-  fars_summarize_years(c(2013, 2014))}

}
