% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Write filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer to specify the year}
}
\value{
A character filename for the specified year
}
\description{
This function generates the filename given a specific year.
}
\examples{
\dontrun{year2013 <- make_filename(2013)}
\dontrun{year2014 <- make_filename(2014)}

\export
}
